requirements.txt
# Core data processing
pandas>=1.5.0
geopandas>=0.12.0
pyproj>=3.4.0

# Database and API
supabase>=1.0.0
sqlalchemy>=1.4.0

# Web scraping and social media
requests>=2.28.0
instaloader>=4.9.0

# Geospatial and mapping
folium>=0.14.0
geojson>=2.5.0

# Development and utilities
click>=8.1.0
python-dotenv>=0.19.0
pathlib2>=2.3.0